integrated rest lib into primary
