--- Manatee.Trello.WebApi/Manatee.Trello.WebApi.csproj
@@ -1,49 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard1.3</TargetFrameworks>
-    <SignAssembly>True</SignAssembly>
-    <AssemblyOriginatorKeyFile>Manatee.Trello.WebApi.snk</AssemblyOriginatorKeyFile>
-    <Authors>gregsdennis</Authors>
-    <Company>Little Crab Solutions</Company>
-    <Copyright>Copyright © 2017 Little Crab Solutions</Copyright>
-    <PackageLicenseUrl>https://github.com/gregsdennis/Manatee.Trello/blob/master/LICENSE.txt</PackageLicenseUrl>
-    <PackageProjectUrl>https://github.com/gregsdennis/Manatee.Trello</PackageProjectUrl>
-    <PackageIconUrl>https://bytebucket.org/gregsdennis/manatee.trello/wiki/Resources/Manatee-Trello-WebApi-Icon.png</PackageIconUrl>
-    <PackageReleaseNotes>Please be sure to read the wiki pages at https://github.com/gregsdennis/Manatee.Trello/wiki.  Feedback and questions are welcome through GitHub, Gitter, or on StackOverflow.
-
-v2.0.3
-Rebuilt for correct versioning in Manatee.Trello v2.2.0.
-
-
-v2.0.2
-BUG FIX: POST with no content throws exception.
-v2.0.1
-BUG FIX: Content format not specified.
-      
-v2.0.0
-BREAKING CHANGE: Frameworks below 4.6.1 are no longer supported.
-Updated to NetStandard1.3</PackageReleaseNotes>
-    <PackageTags>trello rest manatee</PackageTags>
-    <AssemblyVersion>2.0.0.0</AssemblyVersion>
-    <FileVersion>2.0.2.0</FileVersion>
-    <Version>2.0.3</Version>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <TargetFrameworkIdentifier>.NETStandard</TargetFrameworkIdentifier>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netstandard1.3|AnyCPU'">
-    <DocumentationFile>bin\Debug\netstandard1.3\Manatee.Trello.WebApi.xml</DocumentationFile>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netstandard1.3|AnyCPU'">
-    <DocumentationFile>bin\Release\netstandard1.3\Manatee.Trello.WebApi.xml</DocumentationFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Manatee.Trello\Manatee.Trello.csproj" />
-  </ItemGroup>
-
-</Project>
