+++ Manatee.Trello/Rest/DefaultRestClientProvider.cs
@@ -1,0 +1,33 @@
+namespace Manatee.Trello.Rest
+{
+	/// <summary>
+	/// Implements IRestClientProvider for WebApi.
+	/// </summary>
+	public class DefaultRestClientProvider : IRestClientProvider
+	{
+		public static DefaultRestClientProvider Instance { get; } = new DefaultRestClientProvider();
+
+		/// <summary>
+		/// Creates requests for the client.
+		/// </summary>
+		public IRestRequestProvider RequestProvider { get; }
+
+		/// <summary>
+		/// Creates a new instance of the <see cref="DefaultRestClientProvider"/> class.
+		/// </summary>
+		private DefaultRestClientProvider()
+		{
+			RequestProvider = new WebApiRequestProvider();
+		}
+
+		/// <summary>
+		/// Creates an instance of IRestClient.
+		/// </summary>
+		/// <param name="apiBaseUrl">The base URL to be used by the client</param>
+		/// <returns>An instance of IRestClient.</returns>
+		public IRestClient CreateRestClient(string apiBaseUrl)
+		{
+			return new WebApiClient();
+		}
+	}
+}
